Initial refactor [WIP]
