+++ src/archimedes/tree/_struct.py
@@ -45,7 +45,7 @@ fields : Function to retrieve fields of a dataclass
     This is useful for introspection and validation of dataclass instances.
 
 replace : Function to create a new dataclass instance with updated fields
-    For pytree nodes created with @pytree_node, use the .replace() method instead.
+    For pytree nodes created with @struct, use the .replace() method instead.
 """
 
 from __future__ import annotations
@@ -54,23 +54,19 @@ import dataclasses
 import functools
 from collections.abc import Callable
 from dataclasses import InitVar, fields, replace
-from typing import Annotated, Any, Literal, Type, TypeVar, Union
+from typing import Any, TypeVar
 
-from pydantic import BaseModel, ConfigDict, Field
 from typing_extensions import dataclass_transform
 
 from ._registry import register_dataclass
 
 __all__ = [
     "field",
-    "pytree_node",
+    "struct",
     "InitVar",
-    "is_pytree_node",
+    "is_struct",
     "fields",
     "replace",
-    "module",
-    "ModuleConfig",
-    "UnionConfig",
 ]
 
 
@@ -84,10 +80,10 @@ def field(
     **kwargs,
 ) -> dataclasses.Field:
     """
-    Create a field specification with pytree-related metadata.
+    Create a field specification with struct-related metadata.
 
     This function extends :py:func:`dataclasses.field()` with additional metadata to
-    control how fields are treated in pytree operations. Fields can be marked as static
+    control how fields are treated in tree operations. Fields can be marked as static
     (metadata) or dynamic (data). Except for the `static` argument, all other arguments
     are passed directly to :py:func:`dataclasses.field()`; see documentation for the
     :py:mod:`dataclasses` module for details.
@@ -96,7 +92,7 @@ def field(
     ----------
     static : bool, default=False
         If True, the field is treated as static metadata rather than dynamic data.
-        Static fields are preserved during pytree transformations but not included
+        Static fields are preserved during tree transformations but not included
         in the flattened representation.
     metadata : dict, optional
         Additional metadata to include in the field specification. This will be
@@ -116,9 +112,9 @@ def field(
     - To mark configuration parameters that shouldn't change during operations
     - To define default values or constructors for fields
 
-    Static fields are not included when you flatten a pytree or apply transformations
-    like ``map``, but they are preserved in the structure and included when you
-    reconstruct the object.
+    Static fields are not included when you flatten tree-structured data or apply
+    transformations like ``map``, but they are preserved in the structure and included
+    when you reconstruct the object.
 
     Examples
     --------
@@ -126,7 +122,7 @@ def field(
     >>> from archimedes import struct
     >>> import numpy as np
     >>>
-    >>> @struct.pytree_node
+    >>> @arc.struct
     >>> class Vehicle:
     ...     # Dynamic state variables (included in flattening)
     ...     position: np.ndarray
@@ -156,7 +152,7 @@ def field(
 
     See Also
     --------
-    pytree_node : Decorator for creating pytree-compatible classes
+    struct : Decorator for creating pytree-compatible classes
     register_dataclass : Register a dataclass as a pytree node
     """
     f: dataclasses.Field = dataclasses.field(
@@ -167,7 +163,7 @@ def field(
 
 
 @dataclass_transform(field_specifiers=(field,))  # type: ignore[literal-required]
-def pytree_node(cls: T | None = None, **kwargs) -> T | Callable:
+def struct(cls: T | None = None, **kwargs) -> T | Callable:
     """
     Decorator to convert a class into a frozen dataclass registered as a pytree node.
 
@@ -218,17 +214,18 @@ def pytree_node(cls: T | None = None, **kwargs) -> T | Callable:
     Examples
     --------
     >>> import archimedes as arc
+    >>> from archimedes.tree import struct, field
     >>> import numpy as np
     >>>
-    >>> @arc.struct.pytree_node
+    >>> @struct
     >>> class Vehicle:
     ...     # Dynamic state variables (included in transformations)
     ...     position: np.ndarray
     ...     velocity: np.ndarray
     ...
     ...     # Static configuration parameters (preserved during transformations)
-    ...     mass: float = arc.struct.field(static=True, default=1000.0)
-    ...     drag_coef: float = arc.struct.field(static=True, default=0.3)
+    ...     mass: float = field(static=True, default=1000.0)
+    ...     drag_coef: float = field(static=True, default=0.3)
     ...
     ...     def kinetic_energy(self):
     ...         return 0.5 * self.mass * np.sum(self.velocity**2)
@@ -249,7 +246,7 @@ def pytree_node(cls: T | None = None, **kwargs) -> T | Callable:
     >>> print(scaled.mass)        # 1000.0 (unchanged)
     >>>
     >>> # Nested pytree nodes
-    >>> @arc.struct.pytree_node
+    >>> @struct
     >>> class System:
     ...     vehicle1: Vehicle
     ...     vehicle2: Vehicle
@@ -266,9 +263,9 @@ def pytree_node(cls: T | None = None, **kwargs) -> T | Callable:
     field : Define fields with pytree-specific metadata
     module: Decorator for creating functional modules
     """
-    # Support passing arguments to the decorator (e.g. @pytree_node(kw_only=True))
+    # Support passing arguments to the decorator (e.g. @struct(kw_only=True))
     if cls is None:
-        return functools.partial(pytree_node, **kwargs)
+        return functools.partial(struct, **kwargs)
 
     # check if already recognized as a pytree node
     if "_arc_dataclass" in cls.__dict__:
@@ -297,18 +294,18 @@ def pytree_node(cls: T | None = None, **kwargs) -> T | Callable:
 
     register_dataclass(data_cls, data_fields, meta_fields)
 
-    # add a _arc_dataclass flag to distinguish from regular dataclasses
-    data_cls._arc_dataclass = True  # type: ignore[attr-defined]
+    # add a _arc_struct flag to distinguish from regular dataclasses
+    data_cls._arc_struct = True  # type: ignore[attr-defined]
 
     return data_cls  # type: ignore
 
 
-def is_pytree_node(obj: Any) -> bool:
+def is_struct(obj: Any) -> bool:
     """
     Check if an object is a registered pytree node.
 
     This function determines whether an object was created using the
-    :py:func:`pytree_node` decorator, which indicates it has special handling
+    :py:func:`struct` decorator, which indicates it has special handling
     for pytree operations.
 
     Parameters
@@ -331,7 +328,7 @@ def is_pytree_node(obj: Any) -> bool:
     - For debugging pytree-related functionality
 
     This function specifically checks for objects created with the
-    :py:func:`pytree_node` decorator, not built-in pytree containers like lists,
+    :py:func:`struct` decorator, not built-in pytree containers like lists,
     tuples, and dictionaries.
 
     Examples
@@ -339,13 +336,13 @@ def is_pytree_node(obj: Any) -> bool:
     >>> import archimedes as arc
     >>> import numpy as np
     >>>
-    >>> @arc.struct.pytree_node
+    >>> @arc.struct
     >>> class State:
     ...     x: np.ndarray
     ...     v: np.ndarray
     >>>
     >>> state = State(np.zeros(3), np.ones(3))
-    >>> print(arc.struct.is_pytree_node(state))
+    >>> print(arc.tree.is_struct(state))
     True
     >>>
     >>> # Regular dataclass is not a pytree node
@@ -357,243 +354,15 @@ def is_pytree_node(obj: Any) -> bool:
     ...     v: np.ndarray
     >>>
     >>> regular_state = RegularState(np.zeros(3), np.ones(3))
-    >>> print(arc.struct.is_pytree_node(regular_state))
+    >>> print(arc.tree.is_struct(regular_state))
     False
     >>>
     >>> # Built-in containers aren't custom pytree nodes
-    >>> print(arc.struct.is_pytree_node({"x": np.zeros(3)}))
+    >>> print(arc.tree.is_struct({"x": np.zeros(3)}))
     False
 
     See Also
     --------
-    pytree_node : Decorator for creating pytree-compatible classes
+    struct : Decorator for creating tree-compatible classes
     """
-    return hasattr(obj, "_arc_dataclass")
-
-
-@dataclass_transform(field_specifiers=(field,))  # type: ignore[literal-required]
-def module(cls: T | None = None, **kwargs) -> T | Callable:
-    """
-    Decorator to convert a class into a dataclass suitable for modular system design.
-
-    This decorator creates a dataclass that can be used to build modular, composable
-    systems. Unlike :py:func:`pytree_node`, classes decorated with ``@module`` are not
-    automatically registered with the pytree system, making them suitable for
-    organizational and structural components that don't need to participate in
-    tree transformations.  Common use cases include physics models and modular
-    control systems.
-
-    Parameters
-    ----------
-    cls : type, optional
-        The class to convert into a module dataclass
-    **kwargs : dict
-        Additional keyword arguments passed to dataclasses.dataclass().
-        Unlike :py:func:`pytree_node`, ``frozen`` is not set by default.
-
-    Returns
-    -------
-    decorated_class : type
-        The decorated class, now a dataclass marked as a module.
-
-    Notes
-    -----
-    The key difference from :py:func:`pytree_node` is that ``@module`` classes:
-
-    - Are not automatically registered with the pytree system
-    - Are not frozen by default (mutable unless explicitly frozen)
-    - Are intended for structural organization rather than data transformation
-
-    Use ``@module`` when you want dataclass convenience without pytree behavior,
-    and ``@pytree_node`` when you need the object to participate in tree operations
-    like mapping, flattening, and transformations.
-
-    Modules are designed to work well with ModuleConfig classes for managing
-    configurations in large, modular physics models or algorithms.  The combination
-    of these lets you define clear interfaces, protocols, and configurations for
-    complex, hierarchical systems.
-
-    Examples
-    --------
-    >>> import archimedes as arc
-    >>> from archimedes import struct
-    >>> from typing import Protocol
-    >>> import numpy as np
-    >>>
-    >>> class ComponentModel(Protocol):
-    ...     def __call__(self, x: np.ndarray) -> np.ndarray:
-    ...         ...
-    >>>
-    >>> @struct.module
-    >>> class ComponentA:
-    ...     a: float = 1.0
-    ...
-    ...     def __call__(self, x: np.ndarray) -> np.ndarray:
-    ...         return x * self.a
-    >>>
-    >>> @struct.module
-    >>> class ComponentB:
-    ...     b: float = 2.0
-    ...
-    ...     def __call__(self, x: np.ndarray) -> np.ndarray:
-    ...         return x + self.b
-    >>>
-    >>> @struct.module
-    >>> class System:
-    ...     component: ComponentModel
-    ...
-    >>>
-
-    See Also
-    --------
-    pytree_node : Decorator for creating pytree-compatible classes
-    field : Define fields with specific metadata
-    ModuleConfig : Base class for module configuration management
-    """
-    # Support passing arguments to the decorator (e.g. @module(kw_only=True))
-    if cls is None:
-        return functools.partial(module, **kwargs)
-
-    # check if already recognized as a module
-    if "_arc_module" in cls.__dict__:
-        return cls
-
-    data_cls = dataclasses.dataclass(**kwargs)(cls)  # type: ignore
-    data_cls._arc_module = True  # type: ignore[attr-defined]
-    return data_cls  # type: ignore[return-value]
-
-
-class ModuleConfig(BaseModel):
-    """
-    Base class for creating configuration objects with automatic type discrimination.
-
-    This class extends Pydantic's BaseModel to automatically add a ``type`` field
-    based on the class name, enabling type-safe configuration systems with
-    automatic serialization and validation. Subclasses must specify their type
-    using the ``type`` parameter in the class definition.
-
-    Parameters
-    ----------
-    type : str
-        The type identifier for this configuration class, specified in the
-        class definition using ``ModuleConfig, type="typename"``.
-
-    Notes
-    -----
-    The ``type`` field is automatically added to the class and set to the value
-    specified in the class definition. This enables automatic discrimination
-    when working with unions of different configuration types.
-
-    Subclasses are expected to implement a ``build()`` method that constructs
-    the corresponding module instance based on the configuration parameters.
-    This may include any "offline" validation, preprocessing, or data loading
-    that should occur once at initialization time rather than at runtime.
-
-    Key features:
-
-    - Automatic ``type`` field addition and population
-    - Validation and serialization of the fields
-    - Designed to work with :py:class:`UnionConfig` for type discrimination
-
-    Examples
-    --------
-    >>> from typing import Protocol
-    >>> import archimedes as arc
-    >>> from archimedes import struct
-    >>>
-    >>> class GravityModel(Protocol):
-    ...     def __call__(self, position: np.ndarray) -> np.ndarray:
-    ...         ...
-    >>>
-    >>> @struct.module
-    >>> class ConstantGravity:
-    ...     g0: float
-    ...
-    ...     def __call__(self, position: np.ndarray) -> np.ndarray:
-    ...         return np.array([0, 0, self.g0])
-    >>>
-    >>> class ConstantGravityConfig(struct.ModuleConfig, type="constant"):
-    ...     g0: float = 9.81
-    ...
-    ...     def build(self) -> ConstantGravity:
-    ...         return ConstantGravity(self.g0)
-    >>>
-    >>> ConstantGravityConfig(g0=9.81).build()
-    ConstantGravity(g0=9.81)
-    >>>
-    >>> # Another configuration type
-    >>> class PointGravityConfig(struct.ModuleConfig, type="point"):
-    ...     mu: float = 3.986e14  # m^3/s^2
-    ...     RE: float = 6.3781e6  # m
-    ...     lat: float = 0.0  # deg
-    ...     lon: float = 0.0  # deg
-    ...
-    ...     def build(self) -> PointGravity:
-    ...         # Implementation omitted for brevity
-    ...         pass
-    >>>
-    >>> # Create a discriminated union of configuration types
-    >>> GravityConfig = struct.UnionConfig[ConstantGravityConfig, PointGravityConfig]
-
-    See Also
-    --------
-    UnionConfig : Create discriminated unions of ModuleConfig subclasses
-    module : Decorator for creating modular dataclass components
-    """
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    def __init_subclass__(cls, type: str | None = None, **kwargs):
-        super().__init_subclass__(**kwargs)
-        if type is not None:
-            cls.__annotations__ = {"type": Literal[type], **cls.__annotations__}
-            cls.type = type
-
-    # When printing the config, show the class name and fields only but
-    # not the type field
-    def __repr__(self):
-        rep = super().__repr__()
-        if hasattr(self, "type"):
-            return rep.replace(f"type='{self.type}', ", "")
-        return rep
-
-    def build(self):
-        raise NotImplementedError("Subclasses must implement the build() method.")
-
-
-class UnionConfig:
-    """
-    Discriminated union of ModuleConfig subclasses.
-
-    Usage:
-        AnyConfig = UnionConfig[ConfigTypeA, ConfigTypeB]
-
-    Equivalent to:
-        AnyConfig = Annotated[
-            Union[ConfigTypeA, ConfigTypeB],
-            Field(discriminator="type"),
-        ]
-
-    See Also
-    --------
-    ModuleConfig : Base class for module configuration management
-    module : Decorator for creating modular system components
-    """
-
-    def __class_getitem__(cls, item) -> Type:
-        # Handle single type (UnionConfig[OneType])
-        if not isinstance(item, tuple):
-            item = (item,)
-
-        # Validate that all types inherit from ModuleConfig
-        for config_type in item:
-            if not (
-                isinstance(config_type, type) and issubclass(config_type, ModuleConfig)
-            ):
-                raise TypeError(
-                    f"{config_type} must be a subclass of ModuleConfig. "
-                    f"UnionConfig is only for ModuleConfig discriminated unions."
-                )
-
-        # Create the discriminated union
-        return Annotated[Union[item], Field(discriminator="type")]
+    return hasattr(obj, "_arc_struct")
